requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
ollama
python-docx
python-pptx
openpyxl
pillow
pytesseract
PyMuPDF
pydantic-settings